# whatsapp-clone-ui/src/app/services/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatResponse {
  id?: string;
  isRecipientOnline?: boolean;
  lastMessage?: string;
  lastMessageTime?: string;
  name?: string;
  receiversId?: Array<string>;
  senderId?: string;
  unreadCount?: number;
}