# whatsapp-clone-ui/src/app/services/models/group-chat-resource.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface GroupChatResource {
  name: string;
  receiversIds: Array<string>;
  senderId: string;
}