# whatsapp-clone-ui/src/app/services/models/message-resource.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MediaUploadResource } from '../models/media-upload-resource';
export interface MessageResource {
  chatId?: string;
  content?: string;
  mediaResources?: Array<MediaUploadResource>;
  messageType?: 'TEXT' | 'MEDIA' | 'AUDIO';
  senderId?: string;
}