# whatsapp-clone-ui/src/app/services/models/message-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MediaContentResponse } from '../models/media-content-response';
export interface MessageResponse {
  content?: string;
  createdAt?: string;
  id?: number;
  mediaList?: Array<MediaContentResponse>;
  senderId?: string;
  state?: 'SENT' | 'SEEN';
  type?: 'TEXT' | 'MEDIA' | 'AUDIO';
}